requirements.txt
# https://pypi.org/simple
backports.functools-lru-cache==1.5
certifi==2019.6.16
chardet==3.0.4
cheroot==6.5.6
cherrypy==18.1.2
coloredlogs==10.0
humanfriendly==4.18
idna==2.8
jaraco.functools==2.0
more-itertools==7.2.0
pocketsphinx==0.1.15
portend==2.5
pyaudio==0.2.11
python-vlc==3.0.7110
pytz==2019.2
pyyaml==5.1.2
requests==2.31.0
six==1.12.0
tempora==1.14.1
urllib3==1.25.3
webrtcvad==2.0.10
zc.lockfile==2.0
SpeechRecognition==3.10.0
tuya-connector-python==0.1.2
pyjokes==0.6.0
wikipedia==1.4.0
gTTS==2.3.2
paho-mqtt==1.6.1
googletrans==3.1.0a0
geopy==2.4.1
hijri-converter==2.3.1
